+++ models/profit-centre.js
@@ -12,7 +12,7 @@ module.exports = (sequelize, Sequelize) => {
     sapMasterId: {
       allowNull: false,
       unique: true,
-      type: Sequelize.INTEGER(10),
+      type: Sequelize.INTEGER,
     },
     name: {
       allowNull: false,
